requirements.txt
fastapi
uvicorn
python-dotenv
requests
msal
python-docx
azure-identity
azure-ai-agents
pydantic
pyodbc
openai
pypdf
docx2txt
azure-search-documents
azure-core
azure-keyvault-secrets
PyJWT
pdfplumber